refactor(sbb-skiplink-list): fix e2e tests
