+++ src/components/sbb-skiplink-list/sbb-skiplink-list.e2e.ts
@@ -3,20 +3,18 @@ import { html } from 'lit/static-html.js';
 import { waitForLitRender } from '../../global/testing';
 import { SbbSkiplinkList } from './sbb-skiplink-list';
 import { SbbLink } from '../sbb-link';
-import '../sbb-skiplink-list';
+import '../sbb-link';
+import './sbb-skiplink-list';
 
 describe('sbb-skiplink-list', () => {
-  /** NOTE: These are too hard to migrate and are prone to errors :/
-   * consider that the E2EPage is now the 'document' (you should just delete it)
-   * and that the E2EElement equivalent is directly the SbbComponent (e.g. SbbTimeInput) */
   let element: SbbSkiplinkList;
 
   beforeEach(async () => {
     element = await fixture(html`
       <sbb-skiplink-list>
-        <sbb-link href="1">Link 1</sbb-link>
-        <sbb-link href="2">Link 2</sbb-link>
-        <sbb-link href="3">Link 3</sbb-link>
+        <sbb-link href="#" id="link-1">Link 1</sbb-link>
+        <sbb-link href="#" id="link-2">Link 2</sbb-link>
+        <sbb-link href="#" id="link-3">Link 3</sbb-link>
       </sbb-skiplink-list>
       <button id="button">Focus me</button>
     `);
@@ -34,12 +32,12 @@ describe('sbb-skiplink-list', () => {
     expect(listItemLinks[0]).to.have.style('height', '0px');
     expect(listItemLinks[0]).to.have.style('overflow', 'hidden');
 
-    const firstLink: HTMLElement = listItemLinks[0].querySelector('slot');
+    const firstLink: SbbLink = element.querySelector('#sbb-skiplink-list-link-0');
     firstLink.focus();
     expect(listItemLinks[0]).not.to.have.style('height', '0px');
     expect(listItemLinks[0]).to.have.style('overflow', 'visible');
 
-    const secondLink: SbbLink = document.querySelector('#sbb-skiplink-list-link-1');
+    const secondLink: SbbLink = element.querySelector('#sbb-skiplink-list-link-1');
     secondLink.focus();
     expect(listItemLinks[0]).to.have.style('height', '0px');
     expect(listItemLinks[0]).to.have.style('overflow', 'hidden');
@@ -48,18 +46,17 @@ describe('sbb-skiplink-list', () => {
   });
 
   it('should detected later added links', async () => {
-    await fixture(html`<sbb-skiplink-list></sbb-skiplink-list>`);
-    element = document.querySelector('sbb-skiplink-list');
+    element = await fixture(html`<sbb-skiplink-list></sbb-skiplink-list>`);
 
     await waitForLitRender(element);
 
-    document.querySelector('sbb-skiplink-list').innerHTML = `
+    element.innerHTML = `
         <sbb-link href='1'>Link 1</sbb-link>
         <sbb-link href='2'>Link 2</sbb-link>`;
 
     await waitForLitRender(element);
 
-    expect(document.querySelector('sbb-link')).to.have.attribute('id');
-    expect(document.querySelector('sbb-link')).to.have.attribute('slot');
+    expect(element.querySelector('sbb-link')).to.have.attribute('id');
+    expect(element.querySelector('sbb-link')).to.have.attribute('slot');
   });
 });
